pyproject.toml
[project]
name = "learning-daemon-py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pynetdicom>=3.0.4",
    "python-dotenv>=1.2.1",
]
